renamed from li.che to org.swisspush
