+++ src/main/java/org/swisspush/redisques/handler/AddItemHandler.java
@@ -1,25 +1,25 @@
-package li.chee.vertx.redisques.handler;
+package org.swisspush.redisques.handler;
 
 import io.vertx.core.AsyncResult;
-import li.chee.vertx.redisques.RedisQues;
 import io.vertx.core.Handler;
 import io.vertx.core.eventbus.Message;
 import io.vertx.core.json.JsonObject;
+import org.swisspush.redisques.RedisQues;
 
 /**
- * Class PutLock.
+ * Class AddItemHandler.
  *
  * @author baldim
  */
-public class PutLockHandler implements Handler<AsyncResult<String>> {
+public class AddItemHandler implements Handler<AsyncResult<Long>> {
     private Message<JsonObject> event;
 
-    public PutLockHandler(Message<JsonObject> event) {
+    public AddItemHandler(Message<JsonObject> event) {
         this.event = event;
     }
 
     @Override
-    public void handle(AsyncResult<String> reply) {
+    public void handle(AsyncResult<Long> reply) {
         if(reply.succeeded()){
             event.reply(new JsonObject().put(RedisQues.STATUS, RedisQues.OK));
         } else {
