+++ src/main/java/org/swisspush/redisques/handler/GetAllLocksHandler.java
@@ -1,11 +1,11 @@
-package li.chee.vertx.redisques.handler;
+package org.swisspush.redisques.handler;
 
 import io.vertx.core.AsyncResult;
 import io.vertx.core.Handler;
 import io.vertx.core.eventbus.Message;
 import io.vertx.core.json.JsonArray;
 import io.vertx.core.json.JsonObject;
-import li.chee.vertx.redisques.RedisQues;
+import org.swisspush.redisques.RedisQues;
 
 import java.util.List;
