+++ src/main/java/org/swisspush/redisques/handler/ReplaceItemHandler.java
@@ -1,26 +1,26 @@
-package li.chee.vertx.redisques.handler;
+package org.swisspush.redisques.handler;
 
 import io.vertx.core.AsyncResult;
-import li.chee.vertx.redisques.RedisQues;
+import org.swisspush.redisques.RedisQues;
 import io.vertx.core.Handler;
 import io.vertx.core.eventbus.Message;
 import io.vertx.core.json.JsonObject;
 
 /**
- * Class DeleteAllQueueItems.
+ * Class ReplaceItemHandler.
  *
  * @author baldim
  */
-public class DeleteAllQueueItems implements Handler<AsyncResult<Long>> {
+public class ReplaceItemHandler implements Handler<AsyncResult<String>> {
     private Message<JsonObject> event;
 
-    public DeleteAllQueueItems(Message<JsonObject> event) {
+    public ReplaceItemHandler(Message<JsonObject> event) {
         this.event = event;
     }
 
     @Override
-    public void handle(AsyncResult<Long> reply) {
-        if (reply.result() > 0) {
+    public void handle(AsyncResult<String> reply) {
+        if(reply.succeeded()){
             event.reply(new JsonObject().put(RedisQues.STATUS, RedisQues.OK));
         } else {
             event.reply(new JsonObject().put(RedisQues.STATUS, RedisQues.ERROR));
